using int64_t for the counters

+++ asgn3/bubble.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-uint32_t b_moves, b_compares; // counts for moves and comparisons
+int64_t b_moves, b_compares; // counts for moves and comparisons
 
 void bubble_sort(uint32_t *A, uint32_t n);
