+++ asgn3/shell.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-uint32_t s_moves, s_compares; // counts for moves and comparisons
+int64_t s_moves, s_compares; // counts for moves and comparisons
 
 void shell_sort(uint32_t *A, uint32_t n);
